% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_autodraft.R
\name{add_autodraft}
\alias{add_autodraft}
\title{Add an autodrafter to the DataMuster MongoDB database.}
\usage{
add_autodraft(
  RFID,
  mtag,
  property,
  paddock,
  alms,
  direction,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{the RFIDs of the cattle on the autodrafter}

\item{mtag}{the management tags of the cattle on the autodrafter}

\item{property}{the name of the property}

\item{paddock}{the name of the paddock}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password, contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the infrastructure has been successfully added
}
\description{
This function updates the ALMS unit in the Infrastructure and Cattle collections and adds a new document to the AutoDraft collection in the DataMuster MongoDB database. You can only access this function if you have read and write permissions.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
