% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forage.R
\name{add_forage}
\alias{add_forage}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
add_forage(
  property,
  paddock,
  date,
  method,
  samplesize,
  wetwt,
  drywt,
  photo,
  lat = NULL,
  long = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property}

\item{paddock}{the name of the paddock}

\item{date}{the date of the forage assessment}

\item{method}{the assessment method, either 'pasture cut' or 'visual assessment'}

\item{samplesize}{the sample size of the estimation, either 'quadrant' or 'hectares'}

\item{wetwt}{the weight of the pasture as is, prior to drying (kg)}

\item{drywt}{the weight of the pasture after drying (kg)}

\item{photo}{whether or not a photograph of the assessment site was taken, TRUE or FALSE}

\item{lat}{the latitude coordinate of the assessment site, as a decimal}

\item{long}{the longitute coordinate or the assessment site, as a decimal}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function adds individual or groups of cow calving information in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
