% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gps.R
\name{add_gps}
\alias{add_gps}
\title{Add GPS data to the DataMuster MongoDB database}
\usage{
add_gps(RFID, DateTime, lat, long, voltage, username, password)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{DateTime}{a list of timestamps in POSIXct format}

\item{lat}{the latitude of a coordinate point}

\item{long}{the longitude of a coordinate point}

\item{voltage}{the voltage of the device as a number or percentage}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds GPS data to the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
