% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_paddockhistory.R
\name{add_paddockhistory}
\alias{add_paddockhistory}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
add_paddockhistory(
  RFID,
  cattle_id,
  MTag,
  property,
  Paddock,
  currentPaddock,
  dateIN,
  dateOUT = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{MTag}{a list of cattle management tag number/s}

\item{property}{the name of the property}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{calfMTag}{the calf's management tag number}

\item{birthDate}{the cow's date of calving in date format}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function updates individual or groups of cow calving information in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
