% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pregdata.R
\name{add_pregdata}
\alias{add_pregdata}
\title{Update pregnancy information to DataMuster MongoDB database.}
\usage{
add_pregdata(
  RFID = NULL,
  MTag = NULL,
  property,
  date,
  foetalage = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{MTag}{a list of cattle management tag number/s}

\item{property}{the name of the property}

\item{date}{the date of scanning in date format}

\item{foetalage}{the foetal age at scanning}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function adds individual or groups of cow pregnancy information in the Cattle and CalvingData collections in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
