% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stations.R
\name{add_stations}
\alias{add_stations}
\title{Add a station to the DataMuster MongoDB database.}
\usage{
add_stations(
  stationname,
  lat,
  long,
  area = NULL,
  timezone = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{stationname}{the name of the property}

\item{lat}{the latitude of a coordinate point to locate the property}

\item{long}{the longitude of a coordinate point to locate the property}

\item{area}{the area of the staion in hectares, default is 100 ha}

\item{timezone}{the local timezone of the property, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for the list of accepted timezones, default is Australia/Brisbane}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the station has been successfully added
}
\description{
This function adds a station to the DataMuster MongoDB database. You can only access this function if you have read and write permissions.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
