% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_telemetrywarning.R
\name{add_telemetrywarning}
\alias{add_telemetrywarning}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
add_telemetrywarning(
  property,
  assetid,
  filename,
  paddock,
  lastsignal,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the property of the WOW with no telemetry signal}

\item{assetid}{the asset id of the WOW with no telemetry signal}

\item{filename}{the filename of the WOW with no telemetry signal}

\item{paddock}{the paddock of the WOW with no telemetry signal}

\item{lastsignal}{the last signal of the WOW with no telemetry signal}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function adds data for WOWs with no telemetry signals in the last 3 hours to the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
