% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wowraw.R
\name{add_wowraw}
\alias{add_wowraw}
\title{Add Raw WoW data to the WoWRaw MongoDB database}
\usage{
add_wowraw(data, username, password)
}
\arguments{
\item{data}{is a single file of a column of weights including the RFID read that is inserted between the weights when it is read}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds raw WoW data to the WoWRaw MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
