% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnewinfs.R
\name{addnewinfs}
\alias{addnewinfs}
\title{Add new infrastructure}
\usage{
addnewinfs(property, paddock, infstype, assetID, fileout = NULL,
  active = NULL, training = NULL, telemetry = NULL, date = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{property}{name of the property to be added}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
message to say the user has been successfully added
}
\description{
This function adds new infrastructure to the MongoDB database. You can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
