% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnewpaddock.R
\name{addnewpaddock}
\alias{addnewpaddock}
\title{Add a new paddock}
\usage{
addnewpaddock(property, paddockname, filedir, filename, filetype,
  username = NULL, password = NULL)
}
\arguments{
\item{property}{this is the name of the property}

\item{paddockname}{this is the name of the paddock or paddocks}

\item{filedir}{the location of the spatial file}

\item{filename}{the name of the spatial file}

\item{filetype}{the type of spatial file (e.g. shp, kmz)}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the paddock has been successfully updated
}
\description{
This function adds a new paddock or paddocks to the DataMuster database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
