% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnewstation.R
\name{addnewstation}
\alias{addnewstation}
\title{Add new user}
\usage{
addnewstation(stationname, stationshortname, long, lat, PIC = NULL,
  timezone = NULL, username = NULL, password = NULL)
}
\arguments{
\item{stationname}{name of the property to be added}

\item{stationshortname}{a shortened name of the property to be added}

\item{long}{longitude coordinate of a point on the property}

\item{lat}{latitude coordinate of a point on the property}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
message to say the user has been successfully added
}
\description{
This function adds a new user to the MongoDB database. You can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
