% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtemptag.R
\name{addtemptag}
\alias{addtemptag}
\title{Add new tag to the TempTag collection}
\usage{
addtemptag(RFID, date, weight, location, test_tag = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{RFID}{the name of the user to be added to the DataMuster MongoDB Altas server, preferably without a space between the first and last name}

\item{date}{the email address of the user that was used to register via the DataMuster website}

\item{weight}{the MongoDB user access level, either "user" or "admin"}

\item{Location}{a list of the properties the user will be able to access in MongoDB. If access is "admin", the user will be provided access to all existing properties in the database}
}
\value{
This function writes directly to the database. A message will appear to indicate the number of documents that have been successfully added to the MongoDB database
}
\description{
This function adds a new user to the MongoDB database. You can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
