% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALMSuse.R
\name{ALMSuse}
\alias{ALMSuse}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
ALMSuse(property, paddock = NULL, start = NULL, end = NULL)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{paddock}{this is the name of a paddock or list of paddocks as character entries, if no value is entered then all paddocks are loaded}

\item{start}{provide a start date to be returned, this has to be in date format.}

\item{end}{provide a end date to be returned, this has to be in date format.}
}
\value{
a dataframe with a list of the cattle numbers associated with the ALMS and the number of cattle recorded
}
\description{
This function provides a list of cattle
usage of an ALMS unit for a property from MongoDB. Inputs need to be a list of one or more property names and if only one property a paddock name can be included
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
