% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appalmsusenew.R
\name{appalmsusenew}
\alias{appalmsusenew}
\title{Upload ALMS use data}
\usage{
appalmsusenew(property, start = NULL, end = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{start}{provide a start date to be returned, this has to be in date format.}

\item{end}{provide a end date to be returned, this has to be in date format.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the cattle numbers associated with the ALMS and the number of cattle recorded
}
\description{
This function allows ALMS usage data to be added to the DataMuster database
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
