% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appcalvingupload.R
\name{appcalvingupload}
\alias{appcalvingupload}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
appcalvingupload(dalfID, cowID, date, weight, sex, brand, udder,
  frontteats, rearteats, comment, paddock, property, username = NULL,
  password = NULL)
}
\arguments{
\item{cowID}{the management tag number of the dam}

\item{date}{the calving date}

\item{weight}{the birth weight of the calf}

\item{sex}{the sex of the calf}

\item{brand}{the dam's brand}

\item{udder}{the udder score of the dam}

\item{frontteats}{the teat score of the dam's front teats}

\item{rearteats}{the teat score of the dam's rear teats}

\item{comment}{observations made by the stationhand as txt}

\item{paddock}{the paddock the dam calved in}

\item{property}{the property that the dam calved on}

\item{username}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{password}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{calfID}{the management tag number assigned to the calf}
}
\value{
uploads data in a dataframe to the DMMongoDB database
}
\description{
This function allows calving data to be uploaded from the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
