% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeidupload.R
\name{appeidupload}
\alias{appeidupload}
\title{Upload DataMuster EID data to the DataMuster database}
\usage{
appeidupload(RFID, datetime, ALMS, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of the cattle RFID numbers}

\item{datetime}{a list of timestamps in POSIXct format}

\item{ALMS}{a list of ALMS identification codes as character entries}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the data has been successfully uploaded
}
\description{
This function allows DataMuster EID data to be uploaded via the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
