% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptsisessionupload.R
\name{apptsisessionupload}
\alias{apptsisessionupload}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
apptsisessionupload(Tag, EID, date, liveweight, NLIS = NULL,
  pdkfrom = NULL, pdkto = NULL, notes = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{username}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{password}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{RFID}{the RFID number as txt}

\item{datetime}{in correct format}

\item{weight}{in correct format}

\item{ALMS}{in correct format}
}
\value{
uploads data in a dataframe to the DMMongoDB database
}
\description{
This function allows Gallagher TSi session data to be uploaded from the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
