% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appupdateRFID.R
\name{appupdateRFID}
\alias{appupdateRFID}
\title{Updates cattle RFID tag numbers in the DataMuster database}
\usage{
appupdateRFID(RFID, MTag, newRFID, date, username, password)
}
\arguments{
\item{RFID}{a list of the previous cattle RFID numbers}

\item{MTag}{a list of cattle management tag numbers}

\item{newRFID}{a list of the new cattle RFID numbers}

\item{date}{the date that the new RFID tag was applied, in date format. Default is today's date.}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the RFID tag number has been successfully updated
}
\description{
This function allows cattle RFID numbers to be updated in the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
