% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appwowupload.R
\name{appwowupload}
\alias{appwowupload}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
appwowupload(RFID, datetime, weight, ALMS, username, password)
}
\arguments{
\item{RFID}{the RFID number as txt}

\item{datetime}{in correct format}

\item{weight}{in correct format}

\item{ALMS}{in correct format}

\item{username}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{password}{you will need to request access from Lauren O'Connor for a username to write data to the database}
}
\value{
a dataframe with a list of the RFID numbers, associated management tags and current paddocks the cattle are in
}
\description{
This function allows WoW data to be uploaded from the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
