% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcalmsuse.R
\name{calcalmsuse}
\alias{calcalmsuse}
\title{Calculate cattle ALMS use from the DataMuster database}
\usage{
calcalmsuse(
  property,
  start = NULL,
  end = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property}

\item{start}{a start date to be returned in date format, default is "2014-09-01"}

\item{end}{an end date to be returned in date format, default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the cattle numbers associated with the ALMS and the number of cattle recorded
}
\description{
This function calculates individual or groups of cattle ALMS use from the DataMuster database
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
