% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcweeklywts.R
\name{calcweeklywts}
\alias{calcweeklywts}
\title{Retrieve data on cattle ALMS weekly weights from the DataMuster database}
\usage{
calcweeklywts(
  RFID = NULL,
  start = NULL,
  end = NULL,
  values = NULL,
  s.d = NULL,
  minwt = NULL,
  unit = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{start}{a start date to be returned in date format.}

\item{end}{an end date to be returned in date format.}

\item{values}{the minimum number of daily weight values required to calculate an average weekly weight, default is 4}

\item{s.d}{the minimin standard deviation between daily weight values required to calculate an average weekly weight, default is 25}

\item{minwt}{the minimum daily weight (kg) required to be included to calculate an average weekly weight, default is 10}

\item{unit}{the filename of the ALMS unit to search for}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe of cattle RFID numbers and weekly weight statistics
}
\description{
This function calculates average weekly weights from daily ALMS weights for individual or groups of cattle. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
