% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calvingdate.R
\name{calvingdate}
\alias{calvingdate}
\title{Retrieve data on calving from the DataMuster database}
\usage{
calvingdate(RFID, values = NULL, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{values}{the minimum number of calving date values required, default 1}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of RFID numbers and associated calving dates
}
\description{
This function pulls in calving data for individual or groups of cattle. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
