% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cattlesearch.R
\name{cattlesearch}
\alias{cattlesearch}
\title{Retrieve data on cattle from the DataMuster database}
\usage{
cattlesearch(RFID, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers to query}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the RFID numbers and associated information
}
\description{
This function provides a list of cattle RFID numbers with associated general information (category, breed, sex, property, paddock, database status). If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
