% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowGPS.R
\name{cowGPS}
\alias{cowGPS}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
cowGPS(RFID, DateTime, lat, long, voltage, username, password)
}
\arguments{
\item{RFID}{the RFID number as txt}

\item{DateTime}{in correct format}

\item{lat}{in correct format as a number}

\item{long}{in correct format as a number}

\item{voltage}{in correct format as a number or percentage}

\item{username}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{password}{you will need to request access from Lauren O'Connor for a username to write data to the database}
}
\value{
a dataframe with a list of the RFID numbers, associated management tags and current paddocks the cattle are in
}
\description{
This function allows GPS data to be added for an indivdual animal and can be used to export server data into the DataMuster MongoDB data storage and provide live updates
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
