% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailywtsNEW.R
\name{dailywtsNEW}
\alias{dailywtsNEW}
\title{Retrieve data on cattle ALMS daily weights from the DataMuster database}
\usage{
dailywtsNEW(
  RFID,
  start = NULL,
  end = NULL,
  values = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID numbers}

\item{start}{a start date to be returned in date format.}

\item{end}{an end date to be returned in date format}

\item{values}{the minimum number of weight values required, default is 1}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated daily weights and dates. Please note daily ALMS records are not reliable indicators of cattle weights, the weekly average data is much more reliable. However, the daily values can be useful to identify fine scale changes in weight associated with calving events. The daily data is also available for researchers to identify improved weekly averaging algorithms based on the daily data.
}
\description{
This function provides a list of daily weights for individual or groups of cattle for specified periods. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
