% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmaccess.R
\name{dmaccess}
\alias{dmaccess}
\title{Set up access credentials for DataMuster database}
\usage{
dmaccess(user)
}
\arguments{
\item{user}{this is the username as a text string that you have been allocated to all access to the DataMuster MongoDB Atlas server}

\item{password}{you will be prompted to provide your password to be able to acces the MongoDB}
}
\value{
this establishes username and password credentials on the local machine and avoids having to enter them everytime you use the DMMongoDB functions
}
\description{
The function needs to be run first of all. It stores the credentials on the local machine to provide ongoing access to the DMMongoDB.
Please check with Lauren O'Connor \email{lauren.oconnor@datamuster.net.au} to get login details and make sure you run this function before any others as it will be essential
to providing detailed access.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
