% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propseearch.R
\name{DMMongoDB}
\alias{DMMongoDB}
\alias{propsearch}
\title{DataMuster MongoDB Atlas Access}
\usage{
propsearch(property, x, y)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{username}{this is the unername to be able to access the MongDB and is enetered via a prompt}

\item{password}{this is the password to be able to acces the MongoDB}
}
\value{
a dataframe that with a list of the RFID numbers
}
\description{
DataMuster MongoDB Atlas Access
}
