% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcattle.R
\name{findcattle}
\alias{findcattle}
\title{Retrieve cattle information}
\usage{
findcattle(property = NULL, gender = NULL, paddock = NULL,
  status = NULL, values = NULL, username = NULL, password = NULL)
}
\arguments{
\item{property}{property to search for}

\item{gender}{male or female}

\item{paddock}{paddock on property}

\item{status}{the class of cattle either breeding or growing}

\item{values}{a list of MongoDB cattle collection headers that you want returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle with the list of fields defined in the inputs and searched using the search terms.
}
\description{
This function provides a flexible search tool to find cattle based on a flexible number of search terms. It also allows the user to define what values should be returned via a list of fields. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
