% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_almsuse.R
\name{get_almsuse}
\alias{get_almsuse}
\title{Retrieve ALMS use information from the DataMuster MongoDB database.}
\usage{
get_almsuse(
  RFID = NULL,
  location = NULL,
  start = NULL,
  end = NULL,
  timezone = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{location}{the name of the property to search for}

\item{start}{a start date to be returned in date format, default is "2014-09-01"}

\item{end}{an end date to be returned in date format, default is today's date}

\item{timezone}{the local timezone of the property, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for the list of accepted timezones, default is Australia/Brisbane}

\item{fields}{a list of headers from the ALMSUse collection in the DataMuster MongoDB database to be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers with the list of fields defined in the inputs and searched using the search terms
}
\description{
This function provides a search tool to retrieve ALMS use from the ALMSUse collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
