% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calvingdata.R
\name{get_calvingdata}
\alias{get_calvingdata}
\title{Retrieve calving data from the DataMuster MongoDB database.}
\usage{
get_calvingdata(
  RFID = NULL,
  property = NULL,
  start = NULL,
  end = NULL,
  cow_id = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{start}{a start date to be returned in date format, default is "2014-09-01"}

\item{end}{an end date to be returned in date format, default is today's date}

\item{cow_id}{a list of DataMuster database cow identification numbers}

\item{fields}{a list of headers from the StaticWts collection in the DataMuster MongoDB database to be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated static weight records
}
\description{
This function provides a search tool to retrieve calving data from the CalvingData collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
