% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cattlestwthist.R
\name{get_cattlestwthist}
\alias{get_cattlestwthist}
\title{Retrieve cattle static weight history from the DataMuster MongoDB database.}
\usage{
get_cattlestwthist(
  RFID,
  start = NULL,
  end = NULL,
  values = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{start}{a start date to be returned in date format, default is "2014-09-01"}

\item{end}{an end date to be returned in date format, default is today's date}

\item{values}{the minimum number of values required, default is 1}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated static weight
}
\description{
This function provides a search tool to retrieve cattle static weight history from the Cattle collection in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
