% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infrastructure.R
\name{get_infrastructure}
\alias{get_infrastructure}
\title{Retrieve infrastructure information from the DataMuster MongoDB database.}
\usage{
get_infrastructure(
  property = NULL,
  type = NULL,
  active = NULL,
  training = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property to search for}

\item{type}{type of infrastructure}

\item{active}{TRUE or FALSE, if TRUE, filters the data for infrastructure that is currently active}

\item{training}{TRUE or FALSE}

\item{fields}{a list of headers from the Infrastructure collection in the DataMuster MongoDB database to be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated infrastructure statistics
}
\description{
This function provides a search tool to retrieve infrastructure information from the Infrastructure collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
