% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loadrampdata.R
\name{get_loadrampdata}
\alias{get_loadrampdata}
\title{Retrieve load ramp information from the DataMuster MongoDB database.}
\usage{
get_loadrampdata(
  location = NULL,
  start = NULL,
  timezone = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{location}{the name of the property to search for}

\item{start}{a start date and time to be returned in datetime format}

\item{timezone}{the local timezone of the property, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for the list of accepted timezones, default is Australia/Brisbane}

\item{fields}{a list of headers from the LoadRampData collection in the DataMuster MongoDB database to be returned. If not specified, the RFID, weight, timestamp, and location will be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated load ramp information
}
\description{
This function provides a search tool to retrieve load ramp information from the LoadRampData collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
