% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_newusers.R
\name{get_newusers}
\alias{get_newusers}
\title{Retrieve new user information from the DataMuster MongoDB database.}
\usage{
get_newusers(start = NULL, fields = NULL, username = NULL, password = NULL)
}
\arguments{
\item{start}{a start date to be returned in date format, default is “2014-09-01”}

\item{fields}{a list of headers from the NewUser collection in the DataMuster MongoDB database to be returned. If not specified, the login email and date of creation will be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated user information
}
\description{
This function provides a search tool to retrieve new user information from the NewUser collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
