% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paddock_shapefiles.R
\name{get_paddock_shapefiles}
\alias{get_paddock_shapefiles}
\title{Retrieve paddock shapefiles from the DataMuster MongoDB database.}
\usage{
get_paddock_shapefiles(property, filedir, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to add the paddocks}

\item{filedir}{the location you wish to save the spatial files}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the shapefiles have been successfully saved
}
\description{
This function retrieves paddock shapefiles from the DataMuster MongoDB database.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
