% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paddockhistory.R
\name{get_paddockhistory}
\alias{get_paddockhistory}
\title{Retrieve Paddock historical information from the DataMuster MongoDB database.}
\usage{
get_paddockhistory(
  RFID = NULL,
  property = NULL,
  Paddock = NULL,
  currentPaddock = NULL,
  timezone = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{Paddock}{a list of paddock names to search for}

\item{currentPaddock}{the current paddock allocation, TRUE or FALSE}

\item{timezone}{the local timezone of the property, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for the list of accepted timezones}

\item{fields}{a list of headers from the DailyWts collection in the DataMuster MongoDB database to be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{start}{a start date to be returned in date format, default is "2014-09-01"}

\item{end}{an end date to be returned in date format, default is today's date}
}
\value{
a list of cattle RFID numbers and associated Paddock allocation statistics
}
\description{
This function provides a search tool to retrieve historical Paddock information from the PaddockHistory collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
