% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paddocks.R
\name{get_paddocks}
\alias{get_paddocks}
\title{Retrieve cattle information via the DataMuster website app}
\usage{
get_paddocks(property = NULL, fields = NULL, username = NULL, password = NULL)
}
\arguments{
\item{property}{property to search for}

\item{fields}{a list of MongoDB cattle collection headers that you want returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle with the list of fields defined in the inputs and searched using the search terms.
}
\description{
This function provides a flexible search tool to find cattle based on a flexible number of search terms. It also allows the user to define what fields should be returned via a list of fields.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
