% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Retrieve station information from the DataMuster MongoDB database.}
\usage{
get_stations(
  stationname = NULL,
  report = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{stationname}{the name of the property to search for}

\item{report}{users to search for}

\item{fields}{a list of headers from the Stations collection in the DataMuster MongoDB database to be returned. If not specified, the property name, hectares, timezone, longitude, and latitude will be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated station information
}
\description{
This function provides a search tool to retrieve station information from the Stations collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
