% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taggledata.R
\name{get_taggledata}
\alias{get_taggledata}
\title{Retrieve taggle data information from the DataMuster MongoDB database.}
\usage{
get_taggledata(TagNo, Days, username = NULL, password = NULL)
}
\arguments{
\item{TagNo}{the Taggle tag number}

\item{Days}{number of days to go back and retrieve data for}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe that includes the Taggle tag ID plus the meter readings and the water use per hour
}
\description{
This function provides a search tool to retrieve taggle data from the DMWater collection in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
