% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telemetrywarning.R
\name{get_telemetrywarning}
\alias{get_telemetrywarning}
\title{Retrieve telemetry warnings from the DataMuster MongoDB database.}
\usage{
get_telemetrywarning(
  property,
  assetid,
  lastsignal,
  resolved,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the property of the WOW with no telemetry signal}

\item{assetid}{the asset id of the WOW with no telemetry signal}

\item{lastsignal}{the last signal of the WOW with no telemetry signal}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers with the list of fields defined in the inputs and searched using the search terms
}
\description{
This function provides a search tool to retrieve telemetry warnings from the TelemetryWarnings collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
