% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weeklywts.R
\name{get_weeklywts}
\alias{get_weeklywts}
\title{Retrieve weekly weight information from the DataMuster MongoDB database.}
\usage{
get_weeklywts(
  RFID = NULL,
  cattle_id = NULL,
  start = NULL,
  end = NULL,
  minwt = NULL,
  timezone = NULL,
  fields = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{cattle_id}{a list of cattle database identification number/s}

\item{start}{a start date and time to be returned in datetime format, default is “2014-09-01 00:00:00”}

\item{end}{an end date and time to be returned in datetime format, default is today’s date and time}

\item{minwt}{the minimum weight to be returned}

\item{timezone}{the local timezone of the property, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for the list of accepted timezones, default is Australia/Brisbane}

\item{fields}{a list of headers from the WeeklyWts collection in the DataMuster MongoDB database to be returned. If not specified, the RFID, date, weekly weight, sd of the weights, number of weights, and property name will be returned}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated weekly weight statistics
}
\description{
This function provides a search tool to retrieve weekly weight information from the WeeklyWts collection in the DataMuster MongoDB database. It also allows the user to define what fields should be returned. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
