% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wowdata.R
\name{get_wowdata}
\alias{get_wowdata}
\title{Retrieve WoW information from the DataMuster MongoDB database.}
\usage{
get_wowdata(
  RFID = NULL,
  property = NULL,
  minwt = NULL,
  start = NULL,
  end = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{start}{a start date and time to be returned in datetime format, default is “2014-09-01 00:00:00”}

\item{end}{an end date and time to be returned in datetime format, default is today’s date and time}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username.}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password.}
}
\value{
a dataframe of cattle RFID numbers and associated WoW information
}
\description{
This function provides a search tool to retrieve WoW information from the WoWData collection in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
