% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getacceldata.R
\name{getacceldata}
\alias{getacceldata}
\title{Retrieve accelerometer data from the DataMuster MongoDB database}
\usage{
getacceldata(start = NULL, end = NULL, sample = NULL, username, password)
}
\arguments{
\item{start}{provide a start date and time to be returned, this has to be in datetime format.}

\item{end}{provide a end date and time to be returned, this has to be in datetime format.}

\item{sample}{returns a sample of data for the number of minutes specified. The default is to return all data.}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username.}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password.}
}
\value{
a dataframe of x, y and z axis values in g-force (g).
}
\description{
This function pulls in daily weights for individual or groups of cattle for specified periods. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
