% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgpsdata.R
\name{getgpsdata}
\alias{getgpsdata}
\title{Retrieve accelerometer data from the DataMuster MongoDB database}
\usage{
getgpsdata(start = NULL, end = NULL, username, password)
}
\arguments{
\item{start}{provide a start date and time to be returned, this has to be in datetime format.}

\item{end}{provide a end date and time to be returned, this has to be in datetime format.}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username.}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password.}
}
\value{
a dataframe of GPS coordinates that have been recorded during specified time periods.
}
\description{
This function pulls in GPS coordinates for specified time periods. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
