% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getuhfdata.R
\name{getuhfdata}
\alias{getuhfdata}
\title{Retrieve UHF data from the DataMuster MongoDB database}
\usage{
getuhfdata(start = NULL, end = NULL, username = NULL, password = NULL)
}
\arguments{
\item{start}{provide a start date and time to be returned, this has to be in datetime format.}

\item{end}{provide a end date and time to be returned, this has to be in datetime format.}

\item{username}{required for access. Please email \email{info@datamuster.net.au} to acquire a username.}

\item{password}{required for access. Please email \email{info@datamuster.net.au} to acquire a password.}
}
\value{
a dataframe of UHF data that has been recorded during specified time periods.
}
\description{
This function pulls in data for specified time periods. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
