% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infsearch.R
\name{infsearch}
\alias{infsearch}
\title{Retrieves infrastructure information from the DataMuster database}
\usage{
infsearch(
  property = NULL,
  active = NULL,
  infstype = NULL,
  unitname = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe of infrastructure and associated information
}
\description{
This function allows infrastructure information to be retreived from the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
