% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadrampignore.R
\name{loadrampignore}
\alias{loadrampignore}
\title{Ignore a load ramp record}
\usage{
loadrampignore(RFID, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
message to say the record has been successfully updated
}
\description{
This function updates a record in the MongoDB database. You can only access this function if you have read and write permission
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
