% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newuser.R
\name{newuser}
\alias{newuser}
\title{Add new user}
\usage{
newuser(user, email, access, property)
}
\arguments{
<<<<<<< HEAD
\item{user}{the name of the user to be added to the DataMuster MongoDB Altas server, preferably without a space between the first and last name}

\item{email}{the email address of the user}

\item{access}{the MongoDB user access level, either "user" or "admin"}

\item{property}{a list of the properties the user will be able to access in MongoDB}
=======
\item{user}{this is the user name}

\item{email}{provide a user email address}

\item{access}{provide list of properties that they can access}

\item{property}{name of the property that the user is associated with}
>>>>>>> d2e0bb61963201508d9f6b5791c8ce0deda1b371
}
\value{
This function writes directly to the database. A message will appear to indicate the number of documents that have been successfully added to the Mongo DB database
}
\description{
This function adds a new user to the MongoDB you can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
