% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdkhistsearch.R
\name{pdkhistsearch}
\alias{pdkhistsearch}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
pdkhistsearch(property, paddock = NULL, start = NULL, end = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{paddock}{this is the name of a paddock or list of paddocks as character entries, if no value is entered then all paddocks are loaded}

\item{start}{provide a start date to be returned, this has to be in date format.}

\item{end}{provide a end date to be returned, this has to be in date format.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{archives}{if true, the list of animals returned will include those that are no longer on the property}
}
\value{
a dataframe with a list of the RFID numbers and paddock history
}
\description{
This function provides a list of cattle
for a property from MongoDB including life data. Inputs need to be a list of one or more property names and if only one property a paddock name can be included
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
