% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propcattlecheck.R
\name{propcattlecheck}
\alias{propcattlecheck}
\title{Check cattle access to ALMS accross whole property}
\usage{
propcattlecheck(property, days = NULL)
}
\arguments{
\item{property}{this is a list or single property name as a character entry}

\item{days}{the number of days to search over, if no days entered then it only searches for cattle that have crossed the ALMS today}
}
\value{
a dataframe that provides a list including the RFID number, datetime and weights of cattle
}
\description{
This function pulls in daily weights for cattle for specified periods. It searches based on all cattle across a named property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
