% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propseearch.R
\name{propsearch}
\alias{propsearch}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
propsearch(property)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}
}
\value{
a dataframe with a list of the RFID numbers, associated management tags and current paddocks the cattle are in
}
\description{
This function provides a list of cattle
for a property from MongoDB. Inputs need to be a list of one or more property names and if only one property a paddock name can be included
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
