% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactivatecattle.R
\name{reactivatecattle}
\alias{reactivatecattle}
\title{Reactivate cattle on a station in the DataMuster MongoDB database.}
\usage{
reactivatecattle(
  RFID,
  property,
  paddock,
  date = NULL,
  replacevalues = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{paddock}{the name of the paddock to search for}

\item{date}{the date that the animal left the station in date format, default is today's date}

\item{replacevalues}{the new RFID of the animal}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the cattle has been successfully reactivated and that the RFID tag number has been successfully updated
}
\description{
This function reactivates individual or groups of cattle that have previously been removed from a station on the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
