% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorddeath.R
\name{recorddeath}
\alias{recorddeath}
\title{Record a cattle death on the DataMuster MongoDB database.}
\usage{
recorddeath(
  RFID,
  date = NULL,
  cause = NULL,
  property,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{date}{the date that the animal died in date format, default is today's date}

\item{cause}{the cause of death}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the cow has successfully been assigned a dead status
}
\description{
This function removes individual or groups of cattle from a station on the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
