% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFIDcattlecheck.R
\name{RFIDcattlecheck}
\alias{RFIDcattlecheck}
\title{Check cattle access to ALMS using RFID number}
\usage{
RFIDcattlecheck(RFID, days = NULL)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{days}{the number of days to search over, if no days entered then it only searches for cattle that have crossed the ALMS today}
}
\value{
a dataframe that provides a list including the RFID number, datetime and weights of cattle
}
\description{
This function pulls in daily weights for individual or groups of cattle for specified periods. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. Unlike propsearch this function provides a quick tool to check most recent activity and identify when cattle last visited an ALMS. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
