% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFIDsearch.R
\name{RFIDsearch}
\alias{RFIDsearch}
\title{Search for cattle using RFID in the DataMuster MongoDB database.}
\usage{
RFIDsearch(RFID, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the RFID numbers
}
\description{
This function retrieves RFID from the DataMuster MongoDB database.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
