% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationinfo.R
\name{stationinfo}
\alias{stationinfo}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
stationinfo(property = NULL, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe showing the information associated with the property used to search
}
\description{
This function provides a list of cattle
for a property from MongoDB. Inputs need to be a list of one or more property names and if only one property a paddock name can be included
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
