% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taggledata.R
\name{taggledata}
\alias{taggledata}
\title{Taggle water meter data}
\usage{
taggledata(TagNo, Days, username = NULL, password = NULL)
}
\arguments{
\item{TagNo}{this is a Taggle tag number}

\item{Days}{number of days to go back and pull data in}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
A dataframe that includes the Taggle tag ID plus the meter readings and the water use per hour.
}
\description{
This function pulls in weekly weights for individual or groups of cattle for specified periods. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
