% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_almsrating.R
\name{update_almsrating}
\alias{update_almsrating}
\title{Update ALMS rating for paddocks in the DataMuster MongoDB database.}
\usage{
update_almsrating(
  property,
  paddock,
  ALMSrating,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property to search for}

\item{paddock}{the name of the paddock that the ALMS rating will be applied to}

\item{ALMSrating}{the ALMS rating of the paddock}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the cattle have been successfully allocated to an ALMS unit
}
\description{
This function updates allocates individual or groups of cattle to ALMS units in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
