% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_autodraft.R
\name{update_autodraft}
\alias{update_autodraft}
\title{Update autodrafter details or remove animals from the autodrafter in the DataMuster MongoDB database.}
\usage{
update_autodraft(
  RFID,
  active = NULL,
  alms = NULL,
  direction = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of the previous cattle RFID number/s}

\item{active}{whether the autodrafter is still active}

\item{alms}{the new ALMS ID}

\item{direction}{the direction of the autodrafter}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the RFID tag number has been successfully updated
}
\description{
This function updates the autodrafter details or removes individual or groups of cattle in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
