% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_calvingdata.R
\name{update_calvingdata}
\alias{update_calvingdata}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
update_calvingdata(
  RFID = NULL,
  MTag = NULL,
  season,
  property,
  calfMTag = NULL,
  date = NULL,
  multiples = NULL,
  algdev = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID tag number/s}

\item{MTag}{a list of cattle management tag number/s}

\item{season}{the calving season, e.g. 2018/19}

\item{property}{the name of the property to search for}

\item{calfMTag}{the calf's management tag number}

\item{date}{the cow's date of calving in date format}

\item{multiples}{if the cow had twins or not – TRUE or FALSE. Default is FALSE}

\item{algdev}{if the cow was used for the calving algorithm development - TRUE or FALSE. Default is FALSE}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function updates individual or groups of cow calving information in the Cattle and CalvingData Collections in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
