% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cattlecoords.R
\name{update_cattlecoords}
\alias{update_cattlecoords}
\title{Update a paddock coordinates in the DataMuster MongoDB database.}
\usage{
update_cattlecoords(property, paddock = NULL, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search for}

\item{paddock}{the name of the paddock to search for}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a spatialpointsdataframe with a list of the RFID numbers and a number of other data points, associated management tags and current paddocks the cattle are in
}
\description{
This function updates paddock coordinates in the DataMuster MongoDB database. Inputs need to be a list of one or more property names. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
