% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_foetalagedata.R
\name{update_foetalagedata}
\alias{update_foetalagedata}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
update_foetalagedata(
  RFID,
  date = NULL,
  multiples = NULL,
  algdev = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{date}{the calving date of the cattle, in date format}

\item{multiples}{if the cow had twins or not – TRUE or FALSE. Default is FALSE}

\item{algdev}{if the cow was used for the calving algorithm development - TRUE or FALSE. Default is FALSE}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function updates individual or groups of cow calving information in CalvingData collection in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
