% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_paddock.R
\name{update_paddock}
\alias{update_paddock}
\title{Update cattle allocations to paddocks in the DataMuster MongoDB database.}
\usage{
update_paddock(
  RFID,
  property,
  paddock,
  MTag,
  date = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{paddock}{the name of the paddock to allocate cattle to}

\item{MTag}{a list of cattle management tags}

\item{date}{the date the new paddock allocation was applied in date format, default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the paddocks have been successfully updated
}
\description{
This function allocates individual or groups of cattle to paddocks in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, Anita Chang \email{a.chang@cqu.edu.au}
}
