% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pdkrating.R
\name{update_pdkrating}
\alias{update_pdkrating}
\title{Update ALMS, condition, and long term carrying capacity ratings for paddocks in the DataMuster MongoDB database.}
\usage{
update_pdkrating(
  property,
  paddock,
  ALMSrating = NULL,
  condition = NULL,
  LTCC = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{property}{the name of the property to search for}

\item{paddock}{the name of the paddock that the ALMS rating will be applied to}

\item{ALMSrating}{the ALMS rating of the paddock. Default is 0}

\item{condition}{the condition rating of the paddock. Default is A}

\item{LTCC}{the long term carrying capacity of the paddock. Default is 0}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the paddock has been successfully updated.
}
\description{
This function updates ALMS, condition, and long term carrying capacity ratings in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
