% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pregtest.R
\name{update_pregtest}
\alias{update_pregtest}
\title{Update pregnancy diagnosis results in the DataMuster MongoDB database.}
\usage{
update_pregtest(RFID, foetalage, date = NULL, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{foetalage}{a list of the foetal age in weeks}

\item{date}{the date that the pregnancy diagnosis occurred in date format, default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the pregnancy results have been successfully updated
}
\description{
This function updates pregnancy diagnosis results in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
