% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RFID.R
\name{update_RFID}
\alias{update_RFID}
\title{Update RFID tag number in the DataMuster MongoDB database.}
\usage{
update_RFID(RFID, newRFID, date = NULL, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of the previous cattle RFID number/s}

\item{newRFID}{a list of the new cattle RFID number/s}

\item{date}{the date that the new RFID tag was applied in date format, default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the RFID tag number has been successfully updated
}
\description{
This function updates the RFID number for individual or groups of cattle in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
