% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RFID_MTag.R
\name{update_RFID_MTag}
\alias{update_RFID_MTag}
\title{Updates RFID tag using MTag in the DataMuster MongoDB database.}
\usage{
update_RFID_MTag(
  MTag,
  RFID,
  property,
  date = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{MTag}{a list of cattle management tag number/s}

\item{RFID}{a list of cattle RFID number/s}

\item{property}{the name of the property}

\item{date}{the date that the new RFID tag was applied, in date format. Default is today's date}

\item{username}{a username to access the DataMuster database}

\item{password}{a password to access the DataMuster database}
}
\value{
a message that indicates whether or not the RFID tag number has been successfully updated
}
\description{
This function allows cattle RFID numbers to be updated in the DataMuster database via the DataMuster website
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
