% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stweight.R
\name{update_stweight}
\alias{update_stweight}
\title{Update static weight in the DataMuster MongoDB database.}
\usage{
update_stweight(RFID, weight, date = NULL, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{weight}{a list of the cattle weights}

\item{date}{the date that the weight was measured in date format, default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the static weights have been successfully updated
}
\description{
This function updates static weights for individual or groups of cattle in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
