% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatecalvinginfo.R
\name{updatecalvinginfo}
\alias{updatecalvinginfo}
\title{Update cow calving information to DataMuster database}
\usage{
updatecalvinginfo(
  MTag,
  calfMTag,
  property,
  birthDate = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{MTag}{this is a list of the cow management tag numbers}

\item{calfMTag}{the calf's management tag number}

\item{property}{the name of the property to add the cattle}

\item{birthDate}{the cow's date of calving, this has to be in date format}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{RFID}{this is a list of the cow RFID numbers}

\item{calfRFID}{the calf's RFID number}
}
\value{
a message that indicates the RFID tag number has been successfully updated
}
\description{
This function updates individual or groups of cow calving information to the DataMuster database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
