% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatecattleinfo.R
\name{updatecattleinfo}
\alias{updatecattleinfo}
\title{Update cattle information to DataMuster database}
\usage{
updatecattleinfo(
  RFID,
  MTag = NULL,
  category = NULL,
  weaned = NULL,
  breed = NULL,
  brand = NULL,
  rego = NULL,
  horn = NULL,
  colour = NULL,
  sex = NULL,
  desexed = NULL,
  origin = NULL,
  DOB = NULL,
  birthWeight = NULL,
  damRFID = NULL,
  damMTag = NULL,
  sireRFID = NULL,
  sireMTag = NULL,
  preghistDate = NULL,
  foetalage = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{this is a list of the cattle RFID numbers}

\item{MTag}{this is a list of the cattle management tag numbers}

\item{category}{the class of animal (breeding or growing)}

\item{weaned}{the weaned status of the animal. TRUE if cattle are weaned and FALSE if cattle are not weaned}

\item{breed}{the animal's breed}

\item{brand}{the animal's brand}

\item{rego}{the animal's Breedplan registration number}

\item{horn}{the animal's horn status}

\item{colour}{the animal's sex, male or female}

\item{sex}{the animal's sex, male or female}

\item{desexed}{the animal's desexed status, TRUE if castrated/spayed, FALSE if entire}

\item{origin}{the animal's previous location}

\item{DOB}{the animal's date of bith, this has to be in date format}

\item{birthWeight}{the animal's weight at bith}

\item{damRFID}{the dam's RFID number}

\item{damMTag}{the dam's management tag number}

\item{sireRFID}{the sire's RFID number}

\item{sireMTag}{the sire's management tag number}

\item{preghistDate}{the animal's last preg test date}

\item{foetalage}{the animal's foetal age estimate at last preg test}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{property}{the name of the property to add the cattle}

\item{paddock}{the name of the paddock to add the cattle}

\item{date}{provide the date that the animals were added, this has to be in date format. Default is today's date}
}
\value{
a message that indicates the RFID tag number has been successfully updated
}
\description{
This function updates individual or groups of cattle information to the DataMuster database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
