% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatealms.R, R/updatepaddock.R
\name{updatepaddock}
\alias{updatepaddock}
\alias{updatealms}
\title{Update cattle allocations to alms units}
\usage{
updatealms(RFID, property, alms, date = NULL, username = NULL,
  password = NULL)

updatepaddock(RFID, property, paddock, date = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{date}{provide the date that the new paddock allocation was applied, this has to be in date format. Default is today's date.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}

\item{almsid}{this is the asset_id of the alms unti the cattle are to be allocated.}

\item{RFID}{this is a list of cattle RFID numbers}

\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{date}{provide the date that the new paddock allocation was applied, this has to be in date format. Default is today's date.}

\item{addtoherd}{this indicates whether or not there are cattle already allocated to the paddock. use addtoherd = TRUE to ignore.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the RFID tag number has been successfully updated

a message that indicates the RFID tag number has been successfully updated
}
\description{
This function allocates individual or groups of cattle to alms units. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.

This function allocates individual or groups of cattle to paddocks. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}

Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
