% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatepregtest.R
\name{updatepregtest}
\alias{updatepregtest}
\title{Update pregnancy diagnosis results}
\usage{
updatepregtest(RFID, foetalage, date = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{RFID}{this is a list of the cattle RFID numbers}

\item{foetalage}{this is a list of the cattle weights}

\item{date}{provide the date that the weight was measured, this has to be in date format. Default is today's date.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the weights been successfully updated
}
\description{
This function inserts a new static weight for individual or groups of cattle based on a list of RFID values. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
