% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calfdata.R
\name{add_calfdata}
\alias{add_calfdata}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
add_calfdata(
  RFID = NULL,
  MTag = NULL,
  calfRFID = NULL,
  calfMTag,
  property,
  date,
  weight = NULL,
  multiples = NULL,
  sex = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{MTag}{a list of cattle management tag number/s}

\item{calfRFID}{the calf's RFID tag number}

\item{calfMTag}{the calf's management tag number}

\item{property}{the name of the property}

\item{date}{the calving date in date format}

\item{weight}{the calf's birthweight}

\item{multiples}{if the calf is a twin or single, TRUE or FALSE. Default is FALSE}

\item{sex}{the calf's sex}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function adds individual or groups of cow calving information in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{d.swain@cqu.edu.au}, Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}, and Anita Chang \email{a.chang@cqu.edu.au}
}
