% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALMSuse.R
\name{ALMSuse}
\alias{ALMSuse}
\title{Retrieve data on cattle ALMS use from the DataMuster database}
\usage{
ALMSuse(property, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to query}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe of the cattle RFID numbers and ALMS visit count per day (0 = no visit, 1 = visit)
}
\description{
This function provides a list of cattle usage, by date, of ALMS units for a property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
