% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appcalvingupload.R
\name{appcalvingupload}
\alias{appcalvingupload}
\title{Upload calving data to the DataMuster database}
\usage{
appcalvingupload(
  calfID,
  cowID,
  date,
  weight,
  sex,
  brand,
  udder,
  frontteats,
  rearteats,
  comment,
  paddock,
  property,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{calfID}{a list of calf management tag number/s}

\item{cowID}{a list of cow management tag number/s}

\item{date}{a list of the calving dates in date format}

\item{weight}{a list of calf birth weights}

\item{sex}{the calf's sex (male or female)}

\item{brand}{the cow's brand}

\item{udder}{the cow's udder score}

\item{frontteats}{the cow's front teat score}

\item{rearteats}{the cow's rear teat score}

\item{comment}{observations made by the stationhand as a character entry}

\item{paddock}{the name of the calving paddock}

\item{property}{the name of the property}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the data has been successfully uploaded
}
\description{
This function allows calving data to be uploaded via the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
