% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeidupload.R
\name{appeidupload}
\alias{appeidupload}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
appeidupload(RFID, datetime, ALMS, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{the RFID number as txt}

\item{datetime}{in correct format}

\item{ALMS}{in correct format}

\item{username}{you will need to request access from Lauren O'Connor for a username to write data to the database}

\item{password}{you will need to request access from Lauren O'Connor for a username to write data to the database}
}
\value{
uploads data in a dataframe to the DMMongoDB database
}
\description{
This function allows ALMS EID data to be uploaded from the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
