% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptsisessionupload.R
\name{apptsisessionupload}
\alias{apptsisessionupload}
\title{Upload Gallagher TSi session data to the DataMuster database}
\usage{
apptsisessionupload(Tag, EID, date, liveweight, property, NLIS = NULL,
  pdkfrom = NULL, pdkto = NULL, notes = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{Tag}{a list of cattle management tag numbers}

\item{EID}{a list of cattle RFID numbers}

\item{date}{a list of dates in date format}

\item{liveweight}{a list of cattle weights}

\item{property}{the name of the property}

\item{NLIS}{a list of cattle RFID numbers}

\item{pdkfrom}{a list of paddock names the cattle have moved from}

\item{pdkto}{a list of paddock names the cattle have moved to}

\item{notes}{observations made by the stationhand as a character entry}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the data has been successfully uploaded
}
\description{
This function allows Gallagher TSi session data to be uploaded via the DataMuster website
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
