% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appupdateRFID.R
\name{appupdateRFID}
\alias{appupdateRFID}
\title{Update RFID tag number via the DataMuster website app}
\usage{
appupdateRFID(
  RFID,
  MTag,
  property,
  newRFID,
  date,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{this is a list of the previous cattle RFID numbers}

\item{MTag}{this is a list of the cattle management tag numbers}

\item{property}{this is the name of the station}

\item{newRFID}{this is a list of the new cattle RFID numbers}

\item{date}{provide the date that the new RFID tag was applied, this has to be in date format. Default is today's date.}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates the RFID tag number has been successfully updated
}
\description{
This function pulls in daily weights for individual or groups of cattle for specified periods. It searches based on a list of RFID values. It is recommended that you use the propsearch function to find a list of cattle RFID numbers for a particular property. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
