% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailywts.R
\name{dailywts}
\alias{dailywts}
\title{Cattle daily weights}
\usage{
dailywts(RFID, start = NULL, end = NULL)
}
\arguments{
\item{RFID}{this is a list of cattle RFID numbers}

\item{start}{provide a start date to be returned, this has to be in date format.}

\item{end}{provide a end date to be returned, this has to be in date format.}
}
\value{
is a list that includes a list of the RFID numbers that have been searched, the list of cattle stations that are associated with these RFID numbers and
a dataframe for each RFID number that provides date and weekly average weight for each animal
}
\description{
This function pulls in daily weights for individual or groups of cattle for specified periods. If you need assistance please email info@datamuster.net.au to seek help or suggest improvements.
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
