% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMongoDB.R
\docType{package}
\name{DMMongoDB}
\alias{DMMongoDB}
\title{DMMongoDB: A package to access DataMuster}
\description{
The DMMongoDB package provides important functions to more easily access and process DataMuster data. This package is designed to help researchers,
data analysts and DataMuster partners that want to access data and want to run their own analyses or data presentation.
}
\section{DMMongoDB functions}{

The dmaccess function to help store username and password to enable easy database access..
The dailywts function to enable to access individual or groups of cattle daily weaights from the Automated Livestock Management Systems
The weeklywts function to enable to access individual or groups of cattle daily weaights from the Automated Livestock Management Systems
The propsearch function to identify IDs of cattle from indivudal or groups of cattle properties and filter for individual paddocks.
}

\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
