% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cattlepreghist.R
\name{get_cattlepreghist}
\alias{get_cattlepreghist}
\title{Retrieve cattle pregnancy history from the DataMuster MongoDB database.}
\usage{
get_cattlepreghist(RFID, values = NULL, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{values}{the minimum number of values required, default is 1}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a list of cattle RFID numbers and associated cattle pregnancy history
}
\description{
This function provides a search tool to retrieve cattle pregnancy history from the Cattle collection in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
