% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infsearch.R
\name{infsearch}
\alias{infsearch}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
infsearch(property = NULL, active = NULL, infstype = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search the DataMuster MongoDB Atlas server}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the data numbers, associated management tags and current paddocks the cattle are in
}
\description{
This function provides a list of ALMS units with use history for a property from MongoDB. Inputs need to be a list of one or more property names
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
