% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newuser.R
\name{newuser}
\alias{newuser}
\title{Add new user}
\usage{
newuser(user, email, access, property)
}
\arguments{
\item{user}{this is}

\item{email}{provide a}

\item{access}{provide a}

\item{property}{name of the property}
}
\value{
message to say the user has been successfull added
}
\description{
This function adds a new user to the MongoDB you can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
