% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newuser.R
\name{newuser}
\alias{newuser}
\title{Add new user}
\usage{
newuser(user, email, access, property, username = NULL, password = NULL)
}
\arguments{
\item{user}{this is the user name}

\item{email}{provide a user email address}

\item{access}{provide list of properties that they can access}

\item{property}{name of the property that the user is associated with}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
This function writes directly to the database. A message will appear to indicate the number of documents that have been successfully added to the MongoDB database

message to say the user has been successfully added
}
\description{
This function adds a new user to the MongoDB database. You can only access this function if you have read and write permission
}
\author{
Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
