% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFIDsearch.R
\name{RFIDsearch}
\alias{RFIDsearch}
\title{Package with functions to enable easier code to access to DataMuster MongoDB Atlas servers}
\usage{
RFIDsearch(RFID, username = NULL, password = NULL)
}
\arguments{
\item{RFID}{the list of RFID numbers to search the DataMuster MongoDB Atlas server}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a dataframe with a list of the RFID numbers
}
\description{
This function provides details for a list of cattle
RFID numbers from MongoDB. Inputs need to be a list of one or more RFID numbers
}
\author{
Dave Swain \email{dave.swain@datamuster.net.au} and Lauren O'Connor \email{lauren.oconnor@datamuster.net.au}
}
