% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationinfo.R
\name{stationinfo}
\alias{stationinfo}
\title{Retreive property information from the DataMuster database}
\usage{
stationinfo(property = NULL, username = NULL, password = NULL)
}
\arguments{
\item{property}{the name of the property to search the database}

\item{username}{a username to access the DataMuster database, contact Lauren O'Connor for database access}

\item{password}{a password to access the DataMuster database}
}
\value{
a dataframe showing information associated with the property
}
\description{
This function allows property information to be retreived from the DataMuster database via the DataMuster website
}
\author{
Dave Swain \email{d.swain@cqu.edu.au} and Lauren O'Connor \email{l.r.oconnor@cqu.edu.au}
}
