% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_calvingdata.R
\name{update_calvingdata}
\alias{update_calvingdata}
\title{Update cow calving information to DataMuster MongoDB database.}
\usage{
update_calvingdata(
  RFID = NULL,
  MTag = NULL,
  property = NULL,
  foetalage = NULL,
  foetalagedate = NULL,
  estcalvingdate = NULL,
  birthDate = NULL,
  multiples = NULL,
  DoBalgdev = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle management RFID/s}

\item{MTag}{a list of cattle management tag number/s}

\item{property}{the name of the property}

\item{foetalage}{the foetal age at scanning}

\item{foetalagedate}{the date of scanning in date format}

\item{estcalvingdate}{the estimated date of calving at scanning in date format}

\item{birthDate}{the cow's date of calving in date format}

\item{multiples}{if the calf is a twin or single, TRUE or FALSE}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the calving information has been successfully updated
}
\description{
This function updates individual or groups of cow calving information in the DataMuster MongoDB database. If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
