% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RFIDMTag.R
\name{update_RFIDMTag}
\alias{update_RFIDMTag}
\title{Update RFID using MTag in the DataMuster MongoDB database.}
\usage{
update_RFIDMTag(
  MTag,
  newRFID,
  property,
  date,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{MTag}{a list of cattle management tag number/s}

\item{newRFID}{a list of new cattle RFID number/s}

\item{property}{the name of the property to search for}

\item{date}{the date that the new RFID tag was applied, in date format. Default is today's date}

\item{username}{if you don't have a username set up using the dmaccess function you can pass a username, if no value added then the function looks for a value from dmaccess via keyring}

\item{password}{if you include a username you will also need to add a password contact Lauren O'Connor if you don't have access}
}
\value{
a message that indicates that the RFID tags have been successfully updated
}
\description{
This function updates individual or groups of RFID numbers using MTag in the DataMuster MongoDB database (for use when RFID is "xxx xxxxxxxxxxxx"). If you need assistance please email \email{info@datamuster.net.au} to seek help or suggest improvements.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
